Refactor ModalAgree and Folder components: replace setIsOpen prop with onCloseModal for better clarity, implement FoldersProvider for modal management, and streamline folder deletion and renaming logic using Zustand store.

+++ src/features/folder/ui/FoldersProvider/ModalDeleteFolder/useDeleteFolder/index.ts
@@ -1,15 +1,17 @@
 import { revalidateGetDashboard } from "@/api/dashboard/get-dashboard";
 import { deleteFolder } from "@/api/folder/delete-folder";
 import { ModalAgreeOnAgree } from "@/entities/modal-agree";
+import { useFolderStore } from "@/features/folder/model/store";
 import { useCallback } from "react";
 
-export function useDeleteFolder(folderId: string) {
-  return useCallback<ModalAgreeOnAgree>(
-    async ({ closeModal }) => {
+export function useDeleteFolder() {
+  return useCallback<ModalAgreeOnAgree>(async ({ closeModal }) => {
+    const store = useFolderStore.getState();
+    const folderId = store.modalDeleteFolder.folder?.id;
+    if (typeof folderId === "string") {
       await deleteFolder(folderId);
       await closeModal();
       await revalidateGetDashboard();
-    },
-    [folderId],
-  );
+    }
+  }, []);
 }
